+++ packages/common/package.json
@@ -26,15 +26,14 @@
 	"author": "Prosopo Limited",
 	"license": "Apache-2.0",
 	"dependencies": {
-		"@prosopo/captcha": "2.7.22",
 		"@prosopo/locale": "3.1.4",
-		"@prosopo/util-crypto": "13.5.6",
 		"@prosopo/config": "3.1.5",
 		"i18next": "24.1.0",
 		"zod": "3.23.8"
 	},
 	"devDependencies": {
 		"@types/node": "22.10.2",
+		"@prosopo/types": "3.1.4",
 		"@vitest/coverage-v8": "3.0.9",
 		"concurrently": "9.0.1",
 		"del-cli": "6.0.0",
