+++ packages/common/src/index.ts
@@ -12,7 +12,4 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 export * from "./error.js";
-export * from "./array.js";
-export * from "./hash.js";
-export * from "./types.js";
 export * from "./logger.js";
