+++ packages/common/tsconfig.cjs.json
@@ -18,7 +18,7 @@
 			"path": "../locale/tsconfig.cjs.json"
 		},
 		{
-			"path": "../util-crypto/tsconfig.cjs.json"
+			"path": "../types/tsconfig.cjs.json"
 		}
 	]
 }
