+++ packages/types/package.json
@@ -37,9 +37,7 @@
 		"@polkadot/extension-inject": "0.46.9",
 		"@polkadot/types": "12.6.2",
 		"@polkadot/util": "12.6.2",
-		"@prosopo/common": "3.1.4",
 		"@prosopo/locale": "3.1.4",
-		"@prosopo/types": "3.0.8",
 		"@prosopo/util": "3.0.7",
 		"@prosopo/util-crypto": "13.5.6",
 		"@prosopo/config": "3.1.5",
