+++ packages/types/tsconfig.cjs.json
@@ -15,13 +15,7 @@
 			"path": "../../dev/config/tsconfig.cjs.json"
 		},
 		{
-			"path": "../common/tsconfig.cjs.json"
-		},
-		{
 			"path": "../locale/tsconfig.cjs.json"
-		},
-		{
-			"path": "../types/tsconfig.cjs.json"
 		},
 		{
 			"path": "../util/tsconfig.cjs.json"
